3rd Commit

+++ onlinestore/models.py
@@ -10,7 +10,7 @@ class Customer(models.Model):
         return self.name
 class Product(models.Model):
     name=models.CharField(max_length=200,null=True)
-    price=models.FloatField()
+    price=models.DecimalField(max_digits=7,decimal_places=2)
     digital=models.BooleanField(default=False,null=True,blank=True)
     image=models.ImageField(null=True,blank=True, upload_to='images/')
     def __str__(self):
@@ -36,8 +36,6 @@ class Order(models.Model):
         for i in orderitems:
             if i.product.digital==False:
                 shipping=True
-                
-
         return shipping
     @property
     def get_cart_total(self):
